move silence/speak into core_admin_channels
